requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyyaml==6.0.1
httpx==0.25.2
python-dateutil==2.8.2
aiosqlite==0.19.0
playwright>=1.41.0
apscheduler>=3.10.0
# 3-tier architecture additions
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
